Split to context/spec pattern & extend coverage
I know nspec and subspec and so forth exist, but I'm not so familiar with them that they're my go-to test framework.
They're also not as mainstream as I might like, so I'm going with a light layer on top of xunit to achieve the pattern that I prefer.
I'm writing the tests before the implementation, too. Mostly. ;-)