Add standard error checking & handling
I like APIs that treat clients well by responding in a standard way for 400 Bad Request - I prefer "here are all the things that went wrong" vs "400 BR **** you".
So
* added JE.ApiValidation
* added JE.ApiExceptions for exception-logging since I had IoC issues on the way
* added an IOC registrations test since those get gnarly fast
* added logging within tests, so that I can see in the console what is happening, and also a memory target (so I can treat logs as a 1st-class citizen and assert that they happened)

This raises the question of what to actually respond with - it sounds like this might be an API with sensitive data, in which case what I've done is expose too much to the client about what went wrong with their request, since an attacker can infer from error messages how to proceed.

This is a bigger commit than I'd normally do. Usually, I'd do this, diff it back, and slice it into more constituent parts.