swap to storing usage attempts
this app has the word "Password" in it, which suggests that knowing when things are attempted is likely to be valuable for forensics.
It's usually more useful to store a record of attempts over time than it is to store data in a final state - since that means the "why" of the state is chucked away. So I like to give my data structures a chance to store time-series data.
This way, we can add logic to alert if one user or one password has an abnormally high number of OTP errors, since we have the data.

This is also an illustration of
* the benefit of deferring expensive decisions until later. If I'd gone with a relational database up front, I'd now be messing around with schema migrations and all of that stuff, rather than spiking full speed.
* the benefit of choosing to test at the outer-most layer possible - I established my contracts (including error-cases) early, tested at that level, and have now changed an internal with no test churn whatsoever.